Venue Staff may now view and update tickets
Tickets appear in card-style on the venue staff ticket menu. Ticket information is all displayed. 
Added colored border to the top of the card based on ticket status. Yellow = pending, green = completed. Button at the bottom "Mark as Completed" changes the ticket status to Completed in supabase and changes the card look. Button becomes replaced with a "Re-open Ticket" option in the case a ticket has to be reopened.